requirements.txt
matplotlib==3.8.0
PyQt5==5.15.11
torchsummary==1.5.1
opencv-python==4.8.0.74
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu121
tqdm==4.59.0